+++ test/marp-cli.ts
@@ -7,6 +7,7 @@ import * as cli from '../src/cli'
 import { File } from '../src/file'
 import { Converter, ConvertType } from '../src/converter'
 import { CLIError } from '../src/error'
+import { Server } from '../src/server'
 import { ThemeSet } from '../src/theme'
 import { Watcher } from '../src/watcher'
 
@@ -145,6 +146,28 @@ describe('Marp CLI', () => {
         expect(outputFiles).toContain(assetFn('dist/2.html'))
         expect(outputFiles).toContain(assetFn('dist/3.html'))
         expect(outputFiles).toContain(assetFn('dist/subfolder/5.html'))
+      })
+    })
+
+    context('with --server option', () => {
+      it('starts listening server and watcher for passed directory', async () => {
+        const info = jest.spyOn(cli, 'info')
+        info.mockImplementation()
+
+        const serverStart = jest.spyOn(Server.prototype, 'start')
+        serverStart.mockResolvedValue(0)
+
+        await marpCli(['--input-dir', files, '--server'])
+        expect(info).toBeCalledWith(
+          expect.stringContaining('http://localhost:8080/')
+        )
+        expect(serverStart).toBeCalledTimes(1)
+        expect(Watcher.watch).toHaveBeenCalledWith(
+          [files],
+          expect.objectContaining({
+            mode: Watcher.WatchMode.Notify,
+          })
+        )
       })
     })
 
@@ -455,6 +478,38 @@ describe('Marp CLI', () => {
       expect(cliInfo).toHaveBeenCalledWith(
         expect.stringContaining('4 markdowns')
       )
+    })
+
+    context('with --server option', () => {
+      it('treats passed directory as an input directory of the server', async () => {
+        jest.spyOn(cli, 'info').mockImplementation()
+
+        const serverStart = jest.spyOn(Server.prototype, 'start')
+        serverStart.mockResolvedValue(0)
+
+        await marpCli(['--server', assetFn('_files')])
+        expect(serverStart).toBeCalledTimes(1)
+
+        const server: any = serverStart.mock.instances[0]
+        const converter: Converter = server.converter
+
+        expect(converter.options.inputDir).toBe(assetFn('_files'))
+      })
+    })
+  })
+
+  context('with passing multiple files', () => {
+    const baseArgs = [assetFn('_files/1.md'), assetFn('_files/2.mdown')]
+
+    context('with --server option', () => {
+      it('prints error and return error code', async () => {
+        const error = jest.spyOn(console, 'error').mockImplementation()
+
+        expect(await marpCli([...baseArgs, '--server'])).toBe(1)
+        expect(error).toHaveBeenCalledWith(
+          expect.stringContaining('specify just one directory')
+        )
+      })
     })
   })