+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Add server mode provided by `--server` (`-s`) option ([#27](https://github.com/marp-team/marp-cli/pull/27))
 - Add fonts for internationalization to Docker image ([#26](https://github.com/marp-team/marp-cli/pull/26))
 
 ## v0.0.10 - 2018-09-20